+++ CHANGELOG.md
@@ -82,6 +82,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Json Explorer
 - XML Exporter
 - Request Page Size
+- Http response header number-of-pages
 
 ### Changed
 - Rename Conversion Template as Converter Template